Update lab1.md
Changed a few wordings, most significant of which was the very final exercise. I also changed the dashes to letter labels for the mod exercise, because I guarantee someone will think we mean minus sign. Finally, added the other stuff we thought of as a comment, in case an instructor (i.e. me) wants some inspiration later on.